Closes #4980 Refactor Kinsta thirdparty class (#5011)
* Refactored Kinsta Thirdparty class

* Added new events to kinstra 3rd party class

* Added kinsta and its unit tests

* Fixed code standards and added last unit test

* Added Integration test for rocket_clean_kinsta_cache method

* Added all integration tests

* Change comments from CR

* Fixed `composer.json`

* Added deprecated functions
