+++ inc/3rd-party/3rd-party.php
@@ -14,10 +14,6 @@ if ( rocket_is_plugin_active( 'sg-cachepress/sg-cachepress.php' ) ) {
 	require WP_ROCKET_3RD_PARTY_PATH . 'hosting/siteground.php';
 }
 
-if ( isset( $_SERVER['KINSTA_CACHE_ZONE'] ) ) {
-	require WP_ROCKET_3RD_PARTY_PATH . 'hosting/kinsta.php';
-}
-
 if ( defined( 'PL_INSTANCE_REF' ) && class_exists( '\Presslabs\Cache\CacheHandler' ) && file_exists( WP_CONTENT_DIR . '/advanced-cache.php' ) ) {
 	require WP_ROCKET_3RD_PARTY_PATH . 'hosting/presslabs.php';
 }
